+++ discgenius-vuets/src/components/MixList.vue
